chore(general ledger): Minor changes in delete [GCP-11817] (#8656)

+++ app/Http/Controllers/API/FinalReturnIncomeTemplateLinksAPIController.php
@@ -11,6 +11,7 @@ use App\Http\Controllers\AppBaseController;
 use InfyOm\Generator\Criteria\LimitOffsetCriteria;
 use Prettus\Repository\Criteria\RequestCriteria;
 use Response;
+use App\Models\FinalReturnIncomeReports;
 
 /**
  * Class FinalReturnIncomeTemplateLinksController
@@ -318,9 +319,14 @@ class FinalReturnIncomeTemplateLinksAPIController extends AppBaseController
     {
         /** @var FinalReturnIncomeTemplateLinks $finalReturnIncomeTemplateLinks */
         $finalReturnIncomeTemplateLinks = $this->finalReturnIncomeTemplateLinksRepository->findWithoutFail($id);
+        $isTemplateUsed = FinalReturnIncomeReports::isTemplateUsed($finalReturnIncomeTemplateLinks->templateMasterID);
 
         if (empty($finalReturnIncomeTemplateLinks)) {
             return $this->sendError('Final Return Income Template Links not found');
+        }
+
+        if($isTemplateUsed) {
+            return $this->sendError('Template already used in a report and cannot be deleted', 500);
         }
 
         $finalReturnIncomeTemplateLinks->delete();
